core/requirements.txt
altgraph==0.17.4
annotated-types==0.7.0
anyio==4.6.0
asgiref==3.8.1
auto-py-to-exe==2.33.0
beautifulsoup4==4.12.3
bleach==6.1.0
bottle==0.12.25
bottle-websocket==0.2.9
cachetools==5.5.0
certifi==2024.8.30
charset-normalizer==3.3.2
clipboard==0.0.4
colorama==0.4.6
distro==1.9.0
dj-database-url==2.2.0
Django==4.2.16
django-advanced-filters==2.0.0
django-heroku==0.3.1
django-modeladmin-reorder-reborn==0.1.4
django-nested-admin==4.1.1
django-phonenumber-field==8.0.0
django-phonenumbers==1.0.1
django-stubs==5.1.3
django-stubs-ext==5.1.3
Eel==0.16.0
exceptiongroup==1.2.2
future==0.18.3
gevent==22.10.2
gevent-websocket==0.10.1
google==3.0.0
google-api-core==2.19.1
google-api-python-client==2.142.0
google-auth==2.36.0
google-auth-httplib2==0.2.0
google-auth-oauthlib==1.2.1
googleapis-common-protos==1.64.0
greenlet==2.0.2
gspread==6.1.4
gunicorn==23.0.0
h11==0.14.0
httpcore==1.0.6
httplib2==0.22.0
httpx==0.27.2
idna==3.10
importlib_metadata==8.5.0
jiter==0.5.0
Levenshtein==0.26.1
macholib==1.16.3
Markdown==3.7
mypy==1.15.0
mypy-extensions==1.0.0
numpy==2.0.2
oauthlib==3.2.2
openai==1.51.0
packaging==24.2
pandas==2.2.3
pefile==2024.8.26
phonenumbers==8.13.47
picozero==0.4.2
proto-plus==1.24.0
protobuf==5.27.3
psycopg==3.2.3
psycopg2==2.9.9
psycopg2-binary==2.9.9
pyasn1==0.6.1
pyasn1_modules==0.4.1
pycountry==24.6.1
pydantic==2.9.2
pydantic_core==2.23.4
pyinstaller==5.9.0
pyinstaller-hooks-contrib==2023.1
pyparsing==3.2.0
pyperclip==1.9.0
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
python-Levenshtein==0.26.1
python-monkey-business==1.1.0
pytz==2024.2
pywin32-ctypes==0.2.3
RapidFuzz==3.11.0
regex==2024.9.11
requests==2.32.3
requests-oauthlib==2.0.0
rsa==4.9
simplejson==3.19.3
sip==6.10.0
six==1.16.0
sniffio==1.3.1
soupsieve==2.6
sqlparse==0.5.1
tiktoken==0.8.0
tomli==2.2.1
tqdm==4.66.5
types-PyYAML==6.0.12.20241230
typing_extensions==4.12.2
tzdata==2024.2
uritemplate==4.1.1
urllib3==2.2.3
webencodings==0.5.1
whichcraft==0.6.1
whitenoise==6.7.0
zipp==3.20.2
zope.event==4.6
zope.interface==6.0
